gc, 8g, 8l: fix a handful of warnings
8g/cgen.c
        print format type mismatch

8l/asm.c
        resoff set and not used

gc/pgen.c
        misleading comparison INT > 0x80000000

gc/reflect.c
        dalgsym must be static to match forward declaration

gc/subr.c
        assumed_equal set and not used
        hashmem's second argument is not used

gc/walk.c
        duplicated (unreachable) code

R=rsc
CC=golang-dev
https://golang.org/cl/5651079

+++ src/cmd/gc/pgen.c
@@ -124,7 +124,7 @@ compile(Node *fn)
 		print("allocauto: %lld to %lld\n", oldstksize, (vlong)stksize);
 
 	setlineno(curfn);
-	if(stksize+maxarg > (1ULL<<31))
+	if((int64)stksize+maxarg > (1ULL<<31))
 		yyerror("stack frame too large (>2GB)");
 
 	defframe(ptxt);
